adding new dca-callback-class
