+++ src/Resources/contao/languages/en/tl_content.php
@@ -22,9 +22,3 @@ $GLOBALS['TL_LANG']['tl_content']['sliderBreakpoints'] = ['Breakpoints', 'Overri
 $GLOBALS['TL_LANG']['tl_content']['sliderBreakpointsSettings']['breakpoint'] = ['Breakpoint', 'The breakpoint in pixels.'];
 $GLOBALS['TL_LANG']['tl_content']['sliderBreakpointsSettings']['slidesPerView'] = ['Slides per view', 'Number of visible slides at this breakpoint.'];
 $GLOBALS['TL_LANG']['tl_content']['sliderBreakpointsSettings']['spaceBetween'] = ['Space between', 'Space between slides at this breakpoint.'];
-
-$GLOBALS['TL_LANG']['tl_content']['sliderScripts'] = [
-	'Swiper-scripts',
-	'The Swiper-scripts are <strong>not</strong> combined into the assets.',
-	'The Swiper-scripts are combined into the assets.'
-];
